refactor: extract DateSeparator and OperationListItem from OperationsScreen
- Extract DateSeparator component to handle date separators with spending totals
- Extract OperationListItem component to handle individual operation list items
- Update OperationsScreen to use extracted components
- Remove 11 unused styles (date separator and operation item styles)
- Reduce OperationsScreen.js from 1,284 to 1,136 lines (11.5% reduction)
